+++ Model/Methods/VippsMobilePay.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace GingerPay\Payment\Model\Methods;
+
+use GingerPay\Payment\Redefiners\Model\PaymentLibraryRedefiner;
+
+/**
+ * Vipps MobilePay method class
+ */
+class VippsMobilePay extends PaymentLibraryRedefiner
+{
+
+    /** Payment Code */
+    const METHOD_CODE = 'ginger_methods_vippsmobilepay';
+
+    /**
+     * @var string
+     */
+    public $method_code = self::METHOD_CODE;
+
+    /** Platform Method Code */
+    public $platform_code = 'vipps-mobilepay';
+
+    /**
+     * @var string
+     */
+    protected $_code = self::METHOD_CODE;
+}